+++ inst/testme/test-with.DoPar.R
@@ -40,3 +40,15 @@ local({
   stopifnot(foreach::getDoParName() == "doFuture2")
 })
 stopifnot(foreach::getDoParName() == "doSEQ")
+
+message("*** with.DoPar() without 'expr' ...")
+
+# local = FALSE and 'expr' is missing (should error or fail to eval)
+# with.DoPar(registerDoFuture())
+res <- tryCatch({
+  with(registerDoFuture())
+}, error = identity)
+stopifnot(inherits(res, "error"))
+message("Caught expected error when 'expr' is missing: ", res$message)
+
+message("*** with.DoPar() without 'expr' ... DONE")