+++ NEWS.md
@@ -1,3 +1,14 @@
+# Version 1.2.2 (2026-05-21)
+
+## Bug Fixes
+
+ * Attribute `ordering` of `foreach()` argument `.options.future =
+   list(scheduling = ...)` was ignored. Same for `chunk.size`.
+
+ * `foreach()` argument `.options.future = list(label = ...)` was
+   ignored.
+
+
 # Version 1.2.1 (2026-02-19)
 
 ## Bug Fixes
@@ -19,7 +30,7 @@
 ## Bug Fixes
 
  * Attempts to add or ignore globals via `foreach()` argument
-   `.options.globals` were silently ignored.
+   `.options.future = list(globals = ...)` were silently ignored.
  
 
 # Version 1.1.2 (2025-07-14)